services/core_api/requirements.txt
# Core FastAPI
fastapi>=0.108.0,<0.109.0
uvicorn[standard]>=0.25.0,<0.26.0
python-multipart>=0.0.6,<0.1.0

# Database
sqlalchemy>=2.0.23,<3.0.0
asyncpg>=0.29.0,<0.30.0
alembic>=1.13.0,<2.0.0

# Validation
pydantic[email]>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Authentication & Security
pyjwt>=2.8.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Environment
python-dotenv>=1.2.1,<2.0.0
